pyproject.toml
[project]
name = "quiz-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "jinja2>=3.1.6",
    "json-repair>=0.50.0",
    "langchain-community>=0.3.29",
    "langchain-deepseek>=0.1.4",
    "langchain-tavily>=0.2.11",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.6",
    "pyyaml>=6.0.2",
]
